requirements.txt
pyyaml
asyncio_mqtt
numpy